minor changes to ply_handler and corresponding modules renamed
